refactor: consolidate OAuth provider tests into individual files and remove legacy test suite
